Refactor to compile-per-request if necessary with a pool
wasmtime stores (and anything that accesses it) cannot be used cross-thread.
The previous implementation would segfault if a request was stalled long enough
that a new instance was constructed from an existing store.

Instead, what we do now is keep a pool of fastlike.Instance. Each of these
carries a wasm context and can create a new wasm instance to handle a new
request. Once the request is done, the wasm instance is discarded and other
properties of the fastlike Instance are reset and the fastlike Instance is
returned to the pool.

Currently, a spike in requests will create a new fastlike Instance if one is
not available for reuse, and it will be discarded when the request is
completed. This prevents the pool from growing forever.

To mitigate the cost of compiling for requests, we enable the wasmtime compile
cache, which stores a compiled reference to the module in
$XDG_CACHE_HOME/wasmtime.