Switch backends from RoundTrippers to Handlers
Originally, this project used http.Handler for the backend type, but I switched
to http.RoundTripper because I thought the general use-case would be sending
fastlike subrequests to another process.

While I still believe that to be generally true, having backends be Handlers is
simpler for embedders that want to respond in-process, and it's very easy to
create a reverse proxy to send requests upstream that still implements the
`Handler` interface.